fix(#730): break the org-settings-read <-> server/settings.ts cycle
OrgSettings/OrgBranding/TestTagSettings were defined in server/settings.ts (a
"use server" file) and imported type-only by org-settings-read.ts — which
server/settings.ts itself imports real functions from (readOrgSettings,
saveOrgSettings, etc.), closing a direct 2-file cycle plus the last 3 of the
repo's circular-dependency violations (auth.ts / auth-server.ts / org-context.ts
routed through this pair via org-settings-read.ts).

Moved the three interfaces to a new org-settings-types.ts (pure types, no
imports besides OrgSSOSettings/IncrementReset, both leaf modules). Per
CLAUDE.md, a "use server" file must never re-export a type via `export type
{ X }` (Next's server-action transform emits a runtime reference to the
erased name and crashes SSR) — so every consumer that imported these types
from @/server/settings now imports them from @/lib/org-settings-types
directly instead of server/settings.ts re-exporting them.

This is the last of the repo's 19 circular-dependency violations
(POLICY.md R-3.5) — dependency-cruiser now reports zero.
