+++ src/app/(app)/settings/test-and-tag/page.tsx
@@ -14,10 +14,8 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { FormSection } from "@/components/layout/page-layouts";
-import {
-  updateOrganization,
-  type OrgSettings,
-} from "@/server/settings";
+import { updateOrganization } from "@/server/settings";
+import type { OrgSettings } from "@/lib/org-settings-types";
 import {
   getAuditorTokens,
   createAuditorToken,
